// routes/logic/classes/interfaces/ReCreditingInterface.ts
export interface ReCreditingInterface {
    recreditingNumber: string;
    recreditingDate: string;
    recreditingState: number;
    recreditingPosition: string;
    recreditingCreator: string;
    orderNumber: string;
    deliveryNoteNumber: string;
    invoiceNumber: string;
    invoiceDate: string;
}

export interface ReCreditingDataInterface {
    RE_CREDITING_NUMBER: string;
    RE_CREDITING_DATE: string;
    RE_CREDITING_STATE: number;
    RE_CREDITING_POSITIONS: string;
    RE_CREDITING_CREATOR: string;
    ORDERS_NUMBER: string;
    DELIVERY_NOTES_NUMBER: string;
    INVOICES_NUMBER: string;
    INVOICES_DATE: string;
}